pyproject.toml
[project]
name = "drf-finances"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-filter>=25.1",
    "djangorestframework>=3.15.2",
    "djoser>=2.3.1",
    "drf-yasg>=1.21.10",
    "reportlab>=4.3.1",
]
